requirements.txt
streamlit==1.24.1
openai==0.27.8
python-docx==0.8.11
PyMuPDF==1.21.1
supabase==1.0.0
psycopg2-binary==2.9.6
requests==2.31.0
streamlit-lottie>=0.0.5
google-generativeai>=0.4.1
python-dotenv>=1.0.1

